third_party/WebKit/LayoutTests/svg/animations/svglength-animation-unitType-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test change of unit type for SVGLength animation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS rect.width.baseVal.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS rect.width.animVal.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS rect.width.baseVal.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS rect.width.animVal.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS rect.width.baseVal.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS rect.width.animVal.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS rect.width.baseVal.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS rect.width.animVal.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS rect.width.baseVal.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS successfullyParsed is true

TEST COMPLETE
